+++ test/nstemplatetier/nstemplatetier.go
@@ -1,14 +1,40 @@
 package nstemplatetier
 
 import (
+	"fmt"
 	"testing"
 
 	toolchainv1alpha1 "github.com/codeready-toolchain/api/api/v1alpha1"
 	tierutil "github.com/codeready-toolchain/host-operator/controllers/nstemplatetier/util"
+	"github.com/codeready-toolchain/host-operator/pkg/templates/nstemplatetiers"
+	"github.com/codeready-toolchain/toolchain-common/pkg/test"
 
 	"github.com/stretchr/testify/require"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
+
+func NewNSTemplateTier(tierName string, nsTypes ...string) *toolchainv1alpha1.NSTemplateTier {
+	namespaces := make([]toolchainv1alpha1.NSTemplateTierNamespace, len(nsTypes))
+	for i, nsType := range nsTypes {
+		revision := fmt.Sprintf("123abc%d", i+1)
+		namespaces[i] = toolchainv1alpha1.NSTemplateTierNamespace{
+			TemplateRef: nstemplatetiers.NewTierTemplateName(tierName, nsType, revision),
+		}
+	}
+
+	return &toolchainv1alpha1.NSTemplateTier{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: test.HostOperatorNs,
+			Name:      tierName,
+		},
+		Spec: toolchainv1alpha1.NSTemplateTierSpec{
+			Namespaces: namespaces,
+			ClusterResources: &toolchainv1alpha1.NSTemplateTierClusterResources{
+				TemplateRef: nstemplatetiers.NewTierTemplateName(tierName, "clusterresources", "654321b"),
+			},
+		},
+	}
+}
 
 // PreviousBasicTemplates previous templates for the "basic" tier
 var PreviousBasicTemplates = toolchainv1alpha1.NSTemplateTierSpec{
